Implement HTTP and error logging middlewares, update logger configuration to use environment variable for log level, and remove deprecated logger middleware.
